refactor(atomic-hosted-page): refactor common files into atomic-hosted-ui
https://coveord.atlassian.net/browse/KIT-3900
